Replace tile texts with icons
